+++ plugin_tooling/src-tests/com/github/rustdt/tooling/RustBuildOutputParserTest.java
@@ -1,4 +1,3 @@
-package com.github.rustdt.tooling;
 /*******************************************************************************
  * Copyright (c) 2014, 2014 Bruno Medeiros and other Contributors.
  * All rights reserved. This program and the accompanying materials
@@ -9,6 +8,7 @@ package com.github.rustdt.tooling;
  * Contributors:
  *     Bruno Medeiros - initial API and implementation
  *******************************************************************************/
+package com.github.rustdt.tooling;
 
 
 import static melnorme.lang.tooling.data.StatusLevel.ERROR;