+++ documentation/ChangeLog.md
@@ -1,5 +1,8 @@
 ## RustDT release ChangeLog
 
+### next
+ * Added: Code Completion using Racer. (TODO 'src' dir config)
+
 ### 0.1.1
  * Added: Source menu with shift right/left operations.
  * Added: Toggle Comment action (shortcut: Ctrl+/) to Source Menu.